Move common V3 test data to V3.Tests (#697)
