+++ tests/NuGet.Services.V3.Tests/NuGet.Services.V3.Tests.csproj
@@ -40,7 +40,9 @@
     <Compile Include="Registration\RegistrationClientFacts.cs" />
     <Compile Include="Registration\RegistrationUrlBuilderFacts.cs" />
     <Compile Include="Support\Cursor.cs" />
+    <Compile Include="Support\V3Data.cs" />
     <Compile Include="Support\DbSetMockFactory.cs" />
+    <Compile Include="Support\IterTools.cs" />
     <Compile Include="Support\RecordingLogger.cs" />
     <Compile Include="Support\TestCursorStorage.cs" />
     <Compile Include="Support\TestDbAsyncQueryProvider.cs" />
